Added check for large enough hardware
This check was added mainly because AWS CentOS-6 EC2 instances don't
automatically resize their root partition, so even if you request a
larger volume when provisioning the machine, the root filesystem will
still be miserably small.  I decided it's probably a good idea to add
it to all the scripts since errors are often cryptic with no mention
of "No space left on device".